+++ includes/functions.php
@@ -12,6 +12,15 @@ use Pronamic\WordPress\Pay\Payments\Payment;
 use Pronamic\WordPress\Pay\Subscriptions\Subscription;
 
 /**
+ * Pronamic Pay plugin.
+ *
+ * @return \Pronamic\WordPress\Pay\Plugin
+ */
+function pronamic_pay_plugin() {
+	return \Pronamic\WordPress\Pay\Plugin::instance();
+}
+
+/**
  * Get payment by specified post ID.
  *
  * @param int|string $post_id A payment post ID.
@@ -35,7 +44,8 @@ function get_pronamic_payment_by_meta( $meta_key, $meta_value ) {
 
 	$payment = null;
 
-	$db_query = $wpdb->prepare( "
+	$db_query = $wpdb->prepare(
+		"
 		SELECT
 			post_id
 		FROM
@@ -45,7 +55,10 @@ function get_pronamic_payment_by_meta( $meta_key, $meta_value ) {
 				AND
 			meta_value = %s
 			;
-	", $meta_key, $meta_value );
+	",
+		$meta_key,
+		$meta_value
+	);
 
 	$post_id = $wpdb->get_var( $db_query ); // WPCS: unprepared SQL ok, db call ok, cache ok.
 
@@ -72,7 +85,8 @@ function get_pronamic_payments_by_meta( $meta_key, $meta_value ) {
 
 	$payments = array();
 
-	$db_query = $wpdb->prepare( "
+	$db_query = $wpdb->prepare(
+		"
 		SELECT
 			post_id
 		FROM
@@ -84,7 +98,10 @@ function get_pronamic_payments_by_meta( $meta_key, $meta_value ) {
 		ORDER BY
 			meta_id ASC
 			;
-	", $meta_key, $meta_value );
+	",
+		$meta_key,
+		$meta_value
+	);
 
 	$results = $wpdb->get_results( $db_query ); // WPCS: unprepared SQL ok, db call ok, cache ok.
 
@@ -152,7 +169,8 @@ function get_pronamic_subscription_by_meta( $meta_key, $meta_value ) {
 
 	$subscription = null;
 
-	$db_query = $wpdb->prepare( "
+	$db_query = $wpdb->prepare(
+		"
 		SELECT
 			post_id
 		FROM
@@ -162,7 +180,10 @@ function get_pronamic_subscription_by_meta( $meta_key, $meta_value ) {
 				AND
 			meta_value = %s
 			;
-	", $meta_key, $meta_value );
+	",
+		$meta_key,
+		$meta_value
+	);
 
 	$post_id = $wpdb->get_var( $db_query ); // WPCS: unprepared SQL ok, db call ok, cache ok.
 
@@ -185,7 +206,8 @@ function get_pronamic_subscriptions_by_meta( $meta_key, $meta_value ) {
 
 	$subscriptions = array();
 
-	$db_query = $wpdb->prepare( "
+	$db_query = $wpdb->prepare(
+		"
 		SELECT
 			post_id
 		FROM
@@ -197,7 +219,10 @@ function get_pronamic_subscriptions_by_meta( $meta_key, $meta_value ) {
 		ORDER BY
 			meta_id ASC
 			;
-	", $meta_key, $meta_value );
+	",
+		$meta_key,
+		$meta_value
+	);
 
 	$results = $wpdb->get_results( $db_query ); // WPCS: unprepared SQL ok, db call ok, cache ok.