+++ src/Core/Gateway.php
@@ -24,7 +24,7 @@ use WP_Error;
  * Company: Pronamic
  *
  * @author Remco Tolsma
- * @version 2.0.0
+ * @version 2.0.5
  * @since 1.0.0
  */
 abstract class Gateway {
@@ -76,20 +76,6 @@ abstract class Gateway {
 	 * @var int
 	 */
 	private $method;
-
-	/**
-	 * Indiactor if this gateway supports feedback
-	 *
-	 * @var boolean
-	 */
-	private $has_feedback;
-
-	/**
-	 * The mimimum amount this gateway can handle
-	 *
-	 * @var float
-	 */
-	private $amount_minimum;
 
 	/**
 	 * The transaction ID
@@ -231,39 +217,32 @@ abstract class Gateway {
 	}
 
 	/**
-	 * Check if this gateway supports feedback
-	 *
-	 * @return boolean true if gateway supports feedback, false otherwise
-	 */
-	public function has_feedback() {
-		return $this->has_feedback;
-	}
-
-	/**
-	 * Set has feedback
+	 * Set has feedback.
 	 *
 	 * @param boolean $has_feedback Feedback from gateway indicator.
+	 *
+	 * @deprecated 2.0.5 Not in use anymore.
 	 */
 	public function set_has_feedback( $has_feedback ) {
-		$this->has_feedback = $has_feedback;
 	}
 
 	/**
 	 * Set the minimum amount required
 	 *
 	 * @param float $amount Minimum payment amount.
+	 *
+	 * @deprecated 2.0.5 Not in use anymore.
 	 */
 	public function set_amount_minimum( $amount ) {
-		$this->amount_minimum = $amount;
 	}
 
 	/**
 	 * Get iDEAL issuers.
 	 *
-	 * @return array|null
+	 * @return array
 	 */
 	public function get_issuers() {
-		return null;
+		return array();
 	}
 
 	/**